+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New #118: Add `payload` parameter to `SendPaidMedia` method.
 - New #118: Add `prizeStarCount ` field to `GiveawayCreated`, `Giveaway`, `GiveawayWinners` and 
   `ChatBoostSourceGiveaway` types.
+- New #118: Add `isStarGiveaway` field to `GiveawayCompleted` type.
 - New #118: Bump `php-http/multipart-stream-builder` dependency to `^1.4.2` version.
 
 ## 0.4.2 August 22, 2024